+++ src/main/java/cn/daenx/myadmin/system/controller/SysDictDetailController.java
@@ -1,13 +1,13 @@
 package cn.daenx.myadmin.system.controller;
 
+import cn.daenx.myadmin.common.exception.MyException;
 import cn.daenx.myadmin.common.utils.ExcelUtil;
 import cn.daenx.myadmin.common.vo.Result;
-import cn.daenx.myadmin.system.po.SysDict;
 import cn.daenx.myadmin.system.po.SysDictDetail;
 import cn.daenx.myadmin.system.service.SysDictDetailService;
 import cn.daenx.myadmin.system.vo.*;
 import cn.dev33.satoken.annotation.SaCheckPermission;
-import cn.dev33.satoken.annotation.SaIgnore;
+import cn.hutool.core.collection.CollUtil;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import jakarta.annotation.Resource;
 import jakarta.servlet.http.HttpServletResponse;
@@ -103,8 +103,11 @@ public class SysDictDetailController {
      * @return
      */
     @SaCheckPermission("system:dict:remove")
-    @DeleteMapping("/{ids}")
-    public Result remove(@PathVariable String[] ids) {
+    @DeleteMapping()
+    public Result remove(@RequestBody List<String> ids) {
+        if (CollUtil.isEmpty(ids)) {
+            throw new MyException("参数错误");
+        }
         sysDictDetailService.deleteByIds(ids);
         return Result.ok();
     }
