+++ src/main/java/cn/daenx/myadmin/test/controller/TestDataController.java
@@ -1,13 +1,9 @@
 package cn.daenx.myadmin.test.controller;
 
 import cn.daenx.myadmin.common.excel.ExcelResult;
+import cn.daenx.myadmin.common.exception.MyException;
 import cn.daenx.myadmin.common.utils.ExcelUtil;
-import cn.daenx.myadmin.common.utils.MyUtil;
 import cn.daenx.myadmin.common.vo.Result;
-import cn.daenx.myadmin.system.po.SysDict;
-import cn.daenx.myadmin.system.vo.SysDictAddVo;
-import cn.daenx.myadmin.system.vo.SysDictPageVo;
-import cn.daenx.myadmin.system.vo.SysDictUpdVo;
 import cn.daenx.myadmin.test.dto.TestDataPageDto;
 import cn.daenx.myadmin.test.po.TestData;
 import cn.daenx.myadmin.test.service.TestDataService;
@@ -15,8 +11,7 @@ import cn.daenx.myadmin.test.vo.TestDataImportVo;
 import cn.daenx.myadmin.test.vo.TestDataPageVo;
 import cn.daenx.myadmin.test.vo.TestDataUpdVo;
 import cn.dev33.satoken.annotation.SaCheckPermission;
-import cn.dev33.satoken.annotation.SaIgnore;
-import cn.hutool.core.bean.BeanUtil;
+import cn.hutool.core.collection.CollUtil;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import jakarta.annotation.Resource;
 import jakarta.servlet.http.HttpServletResponse;
@@ -119,8 +114,11 @@ public class TestDataController {
      * @return
      */
     @SaCheckPermission("test:data:remove")
-    @DeleteMapping("/{ids}")
-    public Result remove(@PathVariable String[] ids) {
+    @DeleteMapping()
+    public Result remove(@RequestBody List<String> ids) {
+        if (CollUtil.isEmpty(ids)) {
+            throw new MyException("参数错误");
+        }
         testDataService.deleteByIds(ids);
         return Result.ok();
     }
